change id to Long instead of Integer

+++ Wolf_of_Wright_Street/src/main/java/com/wolfofwrightstreet/transactiontracker/model/Transaction.java
@@ -23,7 +23,7 @@ public class Transaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "transaction_id", nullable = false)
-    Integer transactionId;
+    Long transactionId;
 
     @JoinColumn(name = "email_address",  nullable = false)
     @ManyToOne
